+++ lib/aqua.dart
@@ -4,6 +4,7 @@ export 'models/fish.dart' show Fish, FishSpecies, FishSpeciesConfig;
 export 'models/aquatic_creature.dart' show AquaticCreature, CreatureType, CreatureConfig;
 export 'models/aquatic.dart' show Aquatic, AquaticMapper;
 export 'models/ripple.dart' show Ripple;
+export 'models/bubble.dart' show Bubble;
 export 'models/food_pellet.dart' show FoodPellet;
 export 'models/theme.dart' show AquariumThemeData, AquariumThemePreset;